first pass at converting from ActiveRecord to MongoMapper
this is a pretty naive first pass - basically all of the models use
MongoMapper::Document except the models that are direct children of
Patient.  they use EmbeddedDocument.  we'll want to do a more serious
run through all of the model classes to figure out which ones really
need to be standalone and which ones can be embedded.  my guess is
that many should be converted to embedded.