changed HasSelectOptionsExtension from an AR extension to a MM plugin
since MongoMapper works by inclusion rather than subclassing you can't
just open up ActiveRecord::Base and add in stuff that all of the
models will inherit.  that's OK, MM has an explicit plugin API for
that purpose[1].  HasSelectOptionsExtension needed a couple of small
changes to work with MM, and the code to initialize it moved from
environment.rb to an initializer.

[1] http://railstips.org/blog/archives/2010/02/21/mongomapper-07-plugins/