added "key" declarations for models that need them
since mongdb is schemaless, rails can't learn about the models by
looking at the database.  we need to tell MongoMapper about model
fields that it can't otherwise figure out.  this may not be the
complete set but it looks like it's enough to be able to randomize
Patients and save them to the db.