fixed a MongoMapper problem in patient.randomize
this is a very strange problem.  the code used to read:

    self.registration_information ||= RegistrationInformation.new

but in that case registration_information wouldn't be set, or maybe it
would be set but with a proxy that called itself 'nil' but wasn't.  it
wasn't nil because you could call methods on it but they would all
return nil.  i'd like to dig into this but don't have the time for now
- i've got to get something working at least well enough to play
around with.