changed address and telecom into embedded documents
now that i'm getting a chance to play around with mongo and
mongomapper queries i'm learning that "joining" from one collection to
another is very expensive and so you want to use embedded documents
much more than _id links.  when you "join" you have to first look up
the joined-to document and then reference its link in a query, like
so:

>> male = Gender.first 'code'=>'M'
>> males = Patient.all('registration_information.gender_id'=>male._id)

this is a pain.  it's much easier to do:

>> males = Patient.all('registration_information.gender'=>"M")

i'll probably be doing a lot more of this sort of refactoring over the
next couple of days.