another batch of AR->mongomapper changes
reports.rb had a cute polymorphic field thing for the numerator_query
and denominator_query fields.  unfortunately that was pretty dependent
on how AR works under the hood do it didn't convert over to
mongomapper.  i just simplified it so the _query fields are persisted
and there are new accessors called numerator_hash and denominator_hash
that parse the persistent value into yaml.

also removed a lot of the old sql-generation cruft that we won't need
going forward.